Update MMERU ATM MACHINE.cpp

+++ MMERU ATM MACHINE.cpp
@@ -5,9 +5,9 @@ using namespace std;
 int main(){
     int p;
     int pin = 1243;
-    int diposit = 0.0;
-    int balance = 0.0;
-    int withdraw = 0.0;
+    double diposit = 0.0;
+    double balance = 0.0;
+    double withdraw = 0.0;
     int choice;
     int count = 0;
     bool isTrue = true;
